Rename lexible files
