+++ src/games/Lexible/index.ts
@@ -1,2 +1,2 @@
-import LexibleComponent from "./views/LexibleGameComponent";
+import LexibleComponent from "./views/GameComponent";
 export default LexibleComponent;
