+++ src/games/Lexible/models/GameSettings.ts
